requirements.txt
Werkzeug==3.0.0
SQLAlchemy==2.0.19  
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
msal==1.26.0
python-jose==3.3.0
bcrypt==4.1.2
gunicorn==21.2.0
boto3==1.34.34
botocore==1.34.34
watchtower==3.0.1
python-memcached==1.59 